+++ src/zandronum/src/g_shared/shared_hud.cpp
@@ -1042,7 +1042,7 @@ static void DrawTime()
 	const int width  = SmallFont->GetCharWidth('0') * characterCount + 2; // small offset from screen's border
 	const int height = SmallFont->GetHeight();
 
-	DrawHudText(SmallFont, hud_timecolor, timeString, hudwidth - width, height, FRACUNIT);
+	DrawHudText(SmallFont, hud_timecolor, timeString, hudwidth - width, height,(int)( FRACUNIT));
 }
 
 
@@ -1141,7 +1141,7 @@ void DrawHUD()
 		{
 			seconds = level.totaltime / TICRATE;
 			mysnprintf(printstr, countof(printstr), "%02i:%02i:%02i", seconds/3600, (seconds%3600)/60, seconds%60);
-			DrawHudText(SmallFont, hudcolor_ttim, printstr, hudwidth-length, bottom, FRACUNIT);
+			DrawHudText(SmallFont, hudcolor_ttim, printstr, hudwidth-length, bottom,(int)( FRACUNIT));
 			bottom -= fonth;
 		}
 
@@ -1151,14 +1151,14 @@ void DrawHUD()
 			{
 				seconds = level.time /TICRATE;
 				mysnprintf(printstr, countof(printstr), "%02i:%02i:%02i", seconds/3600, (seconds%3600)/60, seconds%60);
-				DrawHudText(SmallFont, hudcolor_time, printstr, hudwidth-length, bottom, FRACUNIT);
+				DrawHudText(SmallFont, hudcolor_time, printstr, hudwidth-length, bottom,(int)( FRACUNIT));
 				bottom -= fonth;
 			}
 
 			// Single level time for hubs
 			seconds= level.maptime /TICRATE;
 			mysnprintf(printstr, countof(printstr), "%02i:%02i:%02i", seconds/3600, (seconds%3600)/60, seconds%60);
-			DrawHudText(SmallFont, hudcolor_ltim, printstr, hudwidth-length, bottom, FRACUNIT);
+			DrawHudText(SmallFont, hudcolor_ltim, printstr, hudwidth-length, bottom,(int)( FRACUNIT));
 		}
 
 		ST_FormatMapName(mapname);
